portfolio started

+++ org.djv.stockresearcher/src/org/djv/stockresearcher/parts/WatchListPart.java
@@ -15,6 +15,7 @@ import org.djv.stockresearcher.model.StockData;
 import org.djv.stockresearcher.widgets.StockTable;
 import org.djv.stockresearcher.widgets.TextProgressBar;
 import org.eclipse.e4.ui.di.Focus;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.SelectionAdapter;
 import org.eclipse.swt.events.SelectionEvent;
@@ -22,6 +23,7 @@ import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.TableItem;
 
 public class WatchListPart implements StockDataChangeListener, WatchListListener{
@@ -34,10 +36,13 @@ public class WatchListPart implements StockDataChangeListener, WatchListListener
 	
 	StockTable table;
 	
+	Shell shell;
+	
 	private StockDB db = StockDB.getInstance();
 	
 	@PostConstruct
 	public void postConstruct(final Composite parent) {
+		this.shell = parent.getShell();
 		createStockTable(parent);
 		db.addStockDataChangeListener(this);
 		db.addWatchListListener(this);
@@ -76,10 +81,13 @@ public class WatchListPart implements StockDataChangeListener, WatchListListener
 			public void widgetSelected(SelectionEvent e) {
 				StockData sd = table.getSelectedStock();
 				if(sd != null){
-					try {
-						db.removeFromWatchList(sd.getStock().getSymbol());
-					} catch (Exception e1) {
-						e1.printStackTrace();
+					boolean result = MessageDialog.openConfirm(shell, "Are you sure?", "Remove \"" + sd.getStock().getSymbol() + "\". Are you sure?");
+					if (result){
+						try {
+							db.removeFromWatchList(sd.getStock().getSymbol());
+						} catch (Exception e1) {
+							e1.printStackTrace();
+						}
 					}
 				}
 			}
